requirements.txt
Flask==3.0.3
matplotlib==3.8.4
numpy==2.1.0
Pillow==10.4.0
torch==2.4.0
transformers==4.44.2
ultralytics==8.2.77
streamlit==1.37.1
torch==2.4.0
torchvision
torchaudio